% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_fit_lines}
\alias{add_fit_lines}
\title{Add Fit Lines}
\usage{
add_fit_lines(x, data)
}
\arguments{
\item{x}{A \code{\link{ggplot}} scatterplot.}

\item{data}{A data.frame with Dayte, estimate, lower and upper values.}
}
\value{
A \code{\link{ggplot}} scatterplot with fit lines.
}
\description{
Adds estimated and lower and upper fit lines to a \code{\link{ggplot}} scatterplot.
}
\seealso{
\code{\link{mwst2}}
}

