% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{add_layer}
\alias{add_layer}
\title{Add Layer to a Map}
\usage{
add_layer(sp, label = NULL, ...)
}
\arguments{
\item{sp}{A sp object.}

\item{label}{NULL or a string of a unique column to use for labelling
centroids.}

\item{...}{Other arguments passed on to ggplot2::layer.}
}
\value{
A list of ggplot2 layers.
}
\description{
Add Layer to a Map
}

