% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_spawners}
\alias{analyse_spawners}
\title{Analyse Spawners}
\usage{
analyse_spawners(x, niter = 10^4)
}
\arguments{
\item{x}{A data.frame with Spawners and Dayte values.}

\item{niter}{A count of the minimum number of MCMC iterations to
perform.}
}
\value{
A jags_analysis object.
}
\description{
Analyses changes in Spawners with Dayte using a Bayesian
spawn timing model.
}
\details{
To view the JAGS model code see \code{\link{spawners_model_code}}
For more information see the manuscript associated with this package.
}
\seealso{
\code{\link{mwst2}}
}

