% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results.R
\name{create_dirs}
\alias{create_dirs}
\title{Create Directories}
\usage{
create_dirs()
}
\description{
Creates directory getOption("mwst2.dir") with the
subdirectories plots, rds and pdf to save plots and results.
}

