% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{crop}
\alias{crop}
\title{Crop}
\usage{
crop(sp, xmin = 1641000, xmax = 1647250, ymin = 613000, ymax = 624250)
}
\arguments{
\item{sp}{A Spatial object.}

\item{xmin}{A number indicating the minimum Easting of the bounding box.}

\item{xmax}{A number indicating the maximum Easting of the bounding box.}

\item{ymin}{A number indicating the minimum Northing of the bounding box.}

\item{ymax}{A number indicating the maximum Northing of the bounding box.}
}
\description{
Crop
}

