% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{gsi_model_code}
\alias{gsi_model_code}
\title{GSI Model Code}
\usage{
gsi_model_code()
}
\value{
A string of the JAGS model code.
}
\description{
Returns a string of the JAGS model code
defining the GSI spawn timing model.
}
\examples{
cat(gsi_model_code())
}
\seealso{
\code{\link{mwst2}}
}

