% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatial.R
\name{map}
\alias{map}
\title{Map Layer}
\usage{
map(sp, label = NULL, ...)
}
\arguments{
\item{sp}{A sp object.}

\item{label}{NULL or a string of a unique column to use for labelling
centroids.}

\item{...}{Other arguments passed on to ggplot2::layer.}
}
\value{
A ggplot2 object.
}
\description{
Map Layer
}

