% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mwst2}
\alias{mwst2}
\alias{mwst2-package}
\title{Mountain Whitefish Spawn Timing R Package}
\description{
\code{mwst2} is the companion R package for the Irvine et al. manuscript
When Do Mountain Whitefish (Prosopium williamsoni) Spawn?
A Comparison of Estimates Based on Gonadosomatic Indices and Spawner and Egg Counts.
}
\details{
For a description of the available demos type \code{demo(package = "mwst2")}.

For a list of the exported objects type \code{ls("package:mwst2")}.
}

