% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gsi}
\alias{plot_gsi}
\title{Plot Gonadal Somatic Index (GSI)}
\usage{
plot_gsi(x)
}
\arguments{
\item{x}{A data.frame with GSI and Dayte values.}
}
\value{
A \code{\link{ggplot}} scatterplot.
}
\description{
Plots GSI by Dayte.
}
\examples{
data(gsi)
plot_gsi(gsi)
}
\seealso{
\code{\link{mwst2}}
}

