% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Residuals}
\usage{
plot_residuals(x)
}
\arguments{
\item{x}{The \code{\link{jags_analysis}} model.}
}
\value{
A \code{\link{ggplot}} residual plot.
}
\description{
Plots standardised residuals by fitted values.
}
\seealso{
\code{\link{mwst2}}
}

