% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spawners}
\alias{plot_spawners}
\title{Plot Spawners}
\usage{
plot_spawners(x)
}
\arguments{
\item{x}{A data.frame with Spawner and Dayte values.}
}
\value{
A \code{\link{ggplot}} scatterplot.
}
\description{
Plots Spawners by Dayte.
}
\examples{
data(spawners)
plot_spawners(spawners)
}
\seealso{
\code{\link{mwst2}}
}

