% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timing}
\alias{plot_timing}
\title{Plot Spawn Timing}
\usage{
plot_timing(x)
}
\arguments{
\item{x}{A data.frame with Event, estimate, lower, upper and Year values.}
}
\value{
A \code{\link{ggplot}} point range plot.
}
\description{
Plots spawn timing predictions (start, peak and end) by Year.
}
\seealso{
\code{\link{mwst2}}
}

