% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_gsi}
\alias{predict_gsi}
\title{Predict Gonadal Somatic Index (GSI)}
\usage{
predict_gsi(x)
}
\arguments{
\item{x}{The \code{\link{jags_analysis}} model.}
}
\value{
A data.frame of the estimates.
}
\description{
Estimates GSI (with lower and upper 95\% CRIs)
by Sex and Dayte.
}
\seealso{
\code{\link{mwst2}}
}

