% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_spawners}
\alias{predict_spawners}
\title{Predict Spawners}
\usage{
predict_spawners(x)
}
\arguments{
\item{x}{The \code{\link{jags_analysis}} model.}
}
\value{
A data.frame of the estimates.
}
\description{
Estimates spawner abundance (with lower and upper 95\% CRIs) by Dayte.
}
\seealso{
\code{\link{mwst2}}
}

