% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_timing}
\alias{predict_timing}
\title{Predict Spawn Timing}
\usage{
predict_timing(x)
}
\arguments{
\item{x}{The \code{\link{jags_analysis}} model.}
}
\value{
A data.frame of the estimates.
}
\description{
Estimates the start (2.5\% of spawning completed), peak (50\% completed)
and end (97.5\% completed) of spawning (all with lower and upper 95\% CRIs).
}
\seealso{
\code{\link{mwst2}}
}

