% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results.R
\name{save_pdf}
\alias{save_pdf}
\title{Save PDF}
\usage{
save_pdf(x, name, width = 6, height = 6, plot_fun = plot, ...)
}
\arguments{
\item{x}{The object to plot.}

\item{name}{A string indicating the file name (without .pdf).}

\item{width}{A number indicating the width in inches.}

\item{height}{A number indicating the height in inches.}

\item{plot_fun}{A function to plot x.}

\item{...}{Additional arguments passed to plot_fun.}
}
\description{
Plots the object to a .pdf file using plot_fun and saves to the pdf directory.
}

