% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot}
\usage{
save_plot(name, width = 2.63, height = 2.63)
}
\arguments{
\item{name}{A string indicating the file name (without .eps).}

\item{width}{A number indicating the width in inches.}

\item{height}{A number indicating the height in inches.}
}
\description{
Saves the last plot displayed to plots as an .eps file.
}

