% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results.R
\name{save_rds}
\alias{save_rds}
\title{Save RDS}
\usage{
save_rds(x, name)
}
\arguments{
\item{x}{The object to save.}

\item{name}{A string indicating the file name (without .rds).}
}
\description{
Saves the object to the rds dir as an .rds file.
}

