% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyse.R
\name{spawners_model_code}
\alias{spawners_model_code}
\title{Spawners Model Code}
\usage{
spawners_model_code()
}
\value{
A string of the JAGS model code.
}
\description{
Returns a string of the JAGS model code
defining the spawner count spawn timing model.
}
\examples{
cat(spawners_model_code())
}
\seealso{
\code{\link{mwst2}}
}

