% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate-results.R
\name{yesno}
\alias{yesno}
\title{Yes or No}
\usage{
yesno(question)
}
\arguments{
\item{question}{a character scalar of the question to ask the user}
}
\value{
A logical scalar indicating whether reponse was yes (TRUE) or no (FALSE).
}
\description{
Asks user a question and returns a TRUE or FALSE to indicate
whether user answered yes or no.
Modified from devtools package internal function so that yes returns TRUE.
}

