% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate-results.R
\name{replicate_results}
\alias{replicate_results}
\title{Replicate Results}
\usage{
replicate_results(mode = "debug", res = 150L, parallel = TRUE,
  ask = TRUE)
}
\arguments{
\item{mode}{A string specifying the mode for the analyses.}

\item{res}{A count indicating the nominal resolution in ppi.}

\item{parallel}{A flag indicating whether to run the chains in parallel.}

\item{ask}{A flag indicating whether to ask before creating results directory.}
}
\description{
Replicates the results by running the package demo. The results are saved
to a folder results in the working directory.
}
\details{
By default (\code{mode = "debug"}) the analyses are run in debug mode
which quickly generates non-convergent results (this is useful for testing
everything works before commiting to the full analyses).

To run the analyses with sufficient iterations for convergence
with an R-hat value of 1.1. set \code{mode = "report"}. To run the
analyses with the same settings as used for the paper set
\code{mode = "paper"}.
}

